requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web framework
streamlit>=1.28.0

# API client
requests>=2.31.0
urllib3>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Scientific computing (for signal processing)
scipy>=1.11.0

# Machine learning (for baseline comparisons)
scikit-learn>=1.3.0
